Adding PNGCrush for PNG optimization
This is accessible via the pngcrush option in the Gruntfile, add
pngcrush: false to stop it.

Adding pngcrusher

viewport size

Adding crusher to main grunt task

Package pngcrush-installer with grunticon, so there are no global dependencies

moving css data into an object of its own

Get stats and render now have grunticon-file being used instead

ALl parts of processFile now are broken apart

Strict mode

Remove crushing from inside of process

Pngcrusher now runs as a node bg process to make this thing work on windows

Adding more things to make this work with windows

Added windows support version of pngcrush installer

Working on making things more windows capable

Updating pnginstaller

Making pngcrush something that is configurable

Better order the file

Better filtering, better error reporting

Better error handling

Spacing

Make sure that if a png is already made, phantom is grabbing the base64 of that

Fixes a problem where the if pngcrush is output to a file that already
exists, it breaks.

Just delete the folder first

+++ tasks/grunticon.js
@@ -15,8 +15,11 @@ module.exports = function( grunt , undefined ) {
 	var uglify = require( 'uglify-js' );
 	var svgo = new (require( 'svgo' ))();
 	var fs = require( 'fs' );
+	var path = require( 'path' );
 
-	var RSVP = require( '../lib/rsvp' );
+	var RSVP = require( path.join( '..', 'lib', 'rsvp' ) );
+	var crushPath = require( 'pngcrush-installer' ).getBinPath();
+	var crusher = require( path.join( '..', 'lib', 'pngcrusher' ) );
 
 	var readDir = function( path ){
 		var promise = new RSVP.Promise();
@@ -81,6 +84,7 @@ module.exports = function( grunt , undefined ) {
 	grunt.registerMultiTask( 'grunticon', 'A mystical CSS icon solution.', function() {
 		var done = this.async();
 
+
 		// just a quick starting message
 		grunt.log.write( "Look, it's a grunticon!\n" );
 
@@ -88,10 +92,10 @@ module.exports = function( grunt , undefined ) {
 		var config = this.options();
 
 		config.files = {
-			loader: __dirname + "/grunticon/static/grunticon.loader.js",
-			banner: __dirname + "/grunticon/static/grunticon.loader.banner.js",
-			preview: __dirname + "/grunticon/static/preview.html",
-			phantom: __dirname + "/grunticon/phantom.js"
+			loader: path.join( __dirname, 'grunticon', 'static', 'grunticon.loader.js'),
+			banner: path.join( __dirname, 'grunticon', 'static', 'grunticon.loader.banner.js'),
+			preview: path.join( __dirname, 'grunticon', 'static', 'preview.html'),
+			phantom: path.join( __dirname,  'grunticon', 'phantom.js')
 		};
 		// fail if config or no src or dest config
 		if( !config || config.src === undefined || config.dest === undefined ){
@@ -100,12 +104,13 @@ module.exports = function( grunt , undefined ) {
 		}
 
 		// make sure src and dest have / at the end
-		if( !config.src.match( /\/$/ ) ){
-				config.src += "/";
+		if( !config.src.match( path.sep + '$' ) ){
+				config.src += path.sep;
 		}
-		if( !config.dest.match( /\/$/ ) ){
-				config.dest += "/";
+		if( !config.dest.match( path.sep + '$' ) ){
+				config.dest += path.sep;
 		}
+
 
 		var asyncCSS = config.files.loader;
 		var asyncCSSBanner = config.files.banner;
@@ -123,15 +128,15 @@ module.exports = function( grunt , undefined ) {
 		var loadersnippet = config.loadersnippet || "grunticon.loader.txt";
 
 		// css references base path for the loader
-		var cssbasepath = config.cssbasepath || "/";
+		var cssbasepath = config.cssbasepath || path.sep;
 
 		var customselectors = JSON.stringify( config.customselectors ) || "{}";
 
 		// folder name (within the output folder) for generated png files
-		var pngfolder = config.pngfolder || "png/";
+		var pngfolder = config.pngfolder || "png" + path.sep;
 		// make sure pngfolder has / at the end
-		if( !pngfolder.match( /\/$/ ) ){
-				pngfolder += "/";
+		if( !pngfolder.match( path.sep + '$' ) ){
+				pngfolder += path.sep;
 		}
 
 		// css class prefix
@@ -154,11 +159,17 @@ module.exports = function( grunt , undefined ) {
 		var colors = JSON.stringify( config.colors || {} );
 
 		var svgosrc = config.src;
-		var tmp = __dirname + "/tmp/";
+		var tmp = path.join( config.dest , 'tmp' , path.sep );
+
+		var compressPNG = config.pngcrush,
+			render, writeCSS;
 
 
 		// create temp directory
 		grunt.log.write( "creating temp directory at:" + tmp );
+		if( grunt.file.exists( tmp ) ){
+			grunt.file.delete( tmp );
+		}
 		grunt.file.mkdir( tmp );
 
 		grunt.log.write( "\nUsing SVGO to optimize.\n" );
@@ -207,8 +218,11 @@ module.exports = function( grunt , undefined ) {
 			// create the output directory
 			grunt.file.mkdir( config.dest );
 
-			// create the output icons directory
-			grunt.file.mkdir( config.dest + pngfolder );
+			// create the tmp output icons directory
+			if( grunt.file.exists( path.join( tmp , pngfolder ))){
+				grunt.file.delete( path.join( tmp, pngfolder ));
+			}
+			grunt.file.mkdir( path.join( tmp , pngfolder ));
 
 
 			// minify the source of the grunticon loader and write that to the output
@@ -219,34 +233,87 @@ module.exports = function( grunt , undefined ) {
 			grunt.file.write( loaderCodeDest, min );
 			grunt.log.write( "\ngrunticon loader file created." );
 
-			// take it to phantomjs to do the rest
-			grunt.log.write( "\ngrunticon now spawning phantomjs..." );
-			var phantomJsPath = require('phantomjs').path;
-			grunt.log.write('(using path: ' + phantomJsPath + ')');
 
 
-			grunt.util.spawn({
-				cmd: phantomJsPath,
-				args: [
-					config.files.phantom,
-					tmp,
-					config.dest,
-					loaderCodeDest,
-					previewHTMLsrc,
-					datasvgcss,
-					datapngcss,
-					urlpngcss,
-					previewhtml,
-					pngfolder,
-					cssprefix,
-					cssbasepath,
-					customselectors,
-					width,
-					height,
-					colors
-				],
-				fallback: ''
-			}, function(err, result, code) {
+			var callPhantom = function( pngf, temp, writeCSS , callback){
+				// take it to phantomjs to do the rest
+				grunt.log.write( "\ngrunticon now spawning phantomjs..." );
+				var phantomJsPath = require('phantomjs').path;
+				grunt.log.write('(using path: ' + phantomJsPath + ')');
+
+				grunt.util.spawn({
+					cmd: phantomJsPath,
+					args: [
+						config.files.phantom,
+						tmp,
+						config.dest,
+						loaderCodeDest,
+						previewHTMLsrc,
+						datasvgcss,
+						datapngcss,
+						urlpngcss,
+						previewhtml,
+						pngf, //path.join( "tmp","png", path.sep),
+						cssprefix,
+						cssbasepath,
+						customselectors,
+						width,
+						height,
+						colors,
+						temp,
+						writeCSS
+					],
+					fallback: ''
+				}, callback );
+			};
+
+			var crush = function( pngfolder ){
+				grunt.log.write( "\ngrunticon now spawning pngcrush..." );
+				grunt.log.write('(using path: ' + crushPath + ')');
+				var render = false;
+				var writeCSS = true;
+
+				if( grunt.file.exists( path.join( config.dest , pngfolder ) ) ){
+					grunt.file.delete( path.join( config.dest, pngfolder ) );
+				}
+				crusher.crush({
+					input: tmp + pngfolder,
+					outputDir:  config.dest + pngfolder,
+					crushPath: crushPath
+				}, function( stdout , stderr ){
+					grunt.verbose.write( stdout );
+					grunt.verbose.write( stderr );
+					callPhantom( pngfolder, render, writeCSS, function(err, result, code) {
+						// TODO boost this up a bit.
+						if( err ){
+							grunt.log.write("\nSomething went wrong with phantomjs...");
+							grunt.fatal( result.stderr );
+							done( false );
+						} else {
+							grunt.log.write( result.stdout );
+							grunt.file.delete( tmp );
+							done();
+						}
+					});
+				});
+			};
+
+			// Get this party started
+			if( compressPNG !== false ){
+				render = true;
+				writeCSS = false;
+			} else {
+				render = true;
+				writeCSS = true;
+			}
+
+			var pngpath;
+			if( render && writeCSS ){
+				pngpath = pngfolder;
+			} else {
+				pngpath = path.join( "tmp", "png", path.sep );
+			}
+			callPhantom( pngpath, render, writeCSS, function(err, result, code) {
 				// TODO boost this up a bit.
 				if( err ){
 					grunt.log.write("\nSomething went wrong with phantomjs...");
@@ -254,10 +321,15 @@ module.exports = function( grunt , undefined ) {
 					done( false );
 				} else {
 					grunt.log.write( result.stdout );
-					grunt.file.delete( tmp );
-					done();
+					if( render && writeCSS ){
+						grunt.file.delete( tmp );
+						done();
+					} else {
+						crush( pngfolder );
+					}
 				}
 			});
+
 		});
 	});
 };
